Consolidate ElectionServlet tests and update POJO documentation
